Call postMessage() twice when sending perf metrics to the parent.
This is a temporary hack to allow us to collect perf data for interactive
examples embedded in beta.developer.{mozilla|allizom}.org. We can't know
if we're embedded in a beta domain page or not, so we just post the
same message to both the beta domain and the regular domain.