Don't specify the origin for postMessage or check origins on message events.
My last PR to this repo modified perf.js to call postMessage()
twice so that data would be sent to the beta domain while we're
supporting that domain. This turns out to be problematic because:

- there were two other postMessage() calls that did not get
  the same treatment, so we still weren't getting the data we
  needed on the beta domain

- sending extra messages causes angry messages in the developer console

- this hack for the beta domain was temporary because the beta
  domain is going away. But we'll have the same problem with the
  wiki domain, and that isn't going away any time soon, so we need
  a more general solution.

So in this PR, I've found all postMessage() calls and message
event handlers and have removed the origin checks. This means that
interactive examples will be able to communicate with the document
in which they are embedded regardless of the origin of that document.
Since there is no sensitive information being exchanged, this should
not be a security issue.